[Feature] formatting codes using eslint

+++ src/components/BookmarkButton.tsx
@@ -1,75 +1,61 @@
-import React, { useContext, useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { FaBookmark } from "react-icons/fa";
 import { FiBookmark } from "react-icons/fi";
-// libs
-import { supabase } from "@/libs/supabase";
-// contexts
+
 import { AuthContext } from "@/contexts/AuthContext";
 import { UtilitiesContext } from "@/contexts/UtilitiesContext";
+import { supabase } from "@/libs/supabase";
 
 type Props = {
   id: string;
-  property?: string;
   type: string;
 };
-export const BookmarkButton = ({ id, property = "default", type }: Props) => {
-  const { user, bookmarks, setBookmarks, upvotes } = useContext(AuthContext);
-  const { baseUrl, setLoginModal } = useContext(UtilitiesContext);
+export const BookmarkButton = ({ id, type }: Props) => {
+  const { bookmarks, setBookmarks, user } = useContext(AuthContext);
+  const { setLoginModal } = useContext(UtilitiesContext);
   const [bookmarked, setBookmarked] = useState<boolean>(false);
 
   const creator_id = type == "creator" && id;
   const collection_slug = type == "collection" && id;
 
-  const addBookmarkHandler = async (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
+  const addBookmarkHandler = async (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     e.preventDefault();
     if (user) {
-      const checkBookmarked = bookmarks.filter(
-        (bookmark) => bookmark.creator_id == id
-      );
+      const checkBookmarked = bookmarks.filter((bookmark) => {
+        return bookmark.creator_id == id;
+      });
       if (checkBookmarked.length == 0) {
-        const { data, error } = await supabase.from("bookmarks").insert([
+        const { data } = await supabase.from("bookmarks").insert([
           {
-            user_id: user.id,
-            creator_id: creator_id,
             collection_slug: collection_slug,
             created_at: new Date(),
+            creator_id: creator_id,
+            user_id: user.id,
           },
         ]);
-        //@ts-ignore
-        setBookmarks([...bookmarks, ...data]);
+        data && setBookmarks([...bookmarks, ...data]);
         setBookmarked(true);
       }
     } else {
       setLoginModal(true);
-      // alert("Please login.");
     }
   };
-  const removeBookmarkHandler = async (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
+  const removeBookmarkHandler = async (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     e.preventDefault();
     if (user) {
       if (type == "creator") {
-        const { data, error } = await supabase
-          .from("bookmarks")
-          .delete()
-          .match({ user_id: user.id, creator_id: creator_id });
-        const removedBookmarks = bookmarks.filter(
-          (bookmark) => bookmark.creator_id != creator_id
-        );
+        await supabase.from("bookmarks").delete().match({ creator_id: creator_id, user_id: user.id });
+        const removedBookmarks = bookmarks.filter((bookmark) => {
+          return bookmark.creator_id != creator_id;
+        });
         setBookmarks(removedBookmarks);
         setBookmarked(false);
       }
       if (type == "collection") {
-        const { data, error } = await supabase
-          .from("bookmark")
-          .delete()
-          .match({ user_id: user.id, collection_slug: collection_slug });
-        const removedBookmarks = bookmarks.filter(
-          (bookmark) => bookmark.collection_slug != id
-        );
+        await supabase.from("bookmark").delete().match({ collection_slug: collection_slug, user_id: user.id });
+        const removedBookmarks = bookmarks.filter((bookmark) => {
+          return bookmark.collection_slug != id;
+        });
         setBookmarks(removedBookmarks);
         setBookmarked(false);
       }
@@ -78,14 +64,14 @@ export const BookmarkButton = ({ id, property = "default", type }: Props) => {
   const checkBookmarked = () => {
     let filterdBookmarkes = [];
     if (type == "creator") {
-      filterdBookmarkes = bookmarks.filter(
-        (bookmark) => bookmark.creator_id === id
-      );
+      filterdBookmarkes = bookmarks.filter((bookmark) => {
+        return bookmark.creator_id === id;
+      });
     }
     if (type == "collection") {
-      filterdBookmarkes = bookmarks.filter(
-        (bookmark) => bookmark.collection_slug === id
-      );
+      filterdBookmarkes = bookmarks.filter((bookmark) => {
+        return bookmark.collection_slug === id;
+      });
     }
     if (filterdBookmarkes.length > 0) {
       setBookmarked(true);
@@ -113,7 +99,6 @@ export const BookmarkButton = ({ id, property = "default", type }: Props) => {
             className=""
           >
             <FaBookmark className="text-xl text-fuchsia-500 opacity-90" />
-            {/* <AiFillHeart className="text-lg text-pink-500 opacity-90" /> */}
           </button>
         ) : (
           <button
@@ -123,7 +108,6 @@ export const BookmarkButton = ({ id, property = "default", type }: Props) => {
             className=""
           >
             <FiBookmark className="text-xl text-white opacity-30" />
-            {/* <AiOutlineHeart className="text-lg text-white opacity-30" /> */}
           </button>
         )}
       </div>
